route/qdisc: adjust API for 64 bit rate/ceil support for htb class
- existing API/ABI must stay unchanged. We cannot change parameter
  types. Ad most we can add new variants that support 64 bit integers.

- rtnl_tc_calc_txtime64() and rtnl_tc_calc_bufsize64() are trivial.
  We should not blow up the public API of libnl for such a thing.
  If the users needs it, they can just reimplement it.

- getters should return an error code. Especially if the return type
  does not support encoding an error there.

- don't add separate rs_rate64/rs_ceil64 field. Instead, extend the
  "rs_rate" field of "struct rtnl_ratespec" to 64 bits. It's internal
  API.

+++ include/netlink/route/tc.h
@@ -100,8 +100,8 @@ extern uint64_t		rtnl_tc_get_stat(struct rtnl_tc *, enum rtnl_tc_stat);
 extern char *		rtnl_tc_stat2str(enum rtnl_tc_stat, char *, size_t);
 extern int		rtnl_tc_str2stat(const char *);
 
-extern int		rtnl_tc_calc_txtime(int, uint64_t);
-extern int		rtnl_tc_calc_bufsize(int, uint64_t);
+extern int		rtnl_tc_calc_txtime(int, int);
+extern int		rtnl_tc_calc_bufsize(int, int);
 extern int		rtnl_tc_calc_cell_log(int);
 
 extern int		rtnl_tc_read_classid_file(void);
